# RollingWindow Size does not capture read lock